Cleaned up dead code and fixed svg import
