+++ client/app/package.json
@@ -33,6 +33,7 @@
     "expo-keep-awake": "~5.0.1",
     "hsluv": "^0.0.3",
     "react": "16.8.3",
+    "react-dom": "^16.8.6",
     "react-native": "https://github.com/expo/react-native/archive/sdk-33.0.0.tar.gz",
     "react-native-elements": "^1.1.0",
     "react-native-emoji": "^1.6.0",
@@ -40,6 +41,7 @@
     "react-native-paper": "^2.15.2",
     "react-native-snap-carousel": "^3.7.5",
     "react-native-svg": "~9.4.0",
+    "react-native-web": "^0.11.4",
     "react-native-webview": "~5.8.1",
     "react-navigation": "^3.11.0",
     "react-redux": "^6.0.1",
@@ -49,7 +51,7 @@
   },
   "devDependencies": {
     "babel-eslint": "^10.0.2",
-    "babel-preset-expo": "^5.0.0",
+    "babel-preset-expo": "^5.1.1",
     "eslint": "^6.0.1",
     "eslint-config-universe": "^1.0.7",
     "expo-cli": "^2.21.2",
