+++ client/app/src/screens/ColorPicker.js
@@ -2,29 +2,26 @@ import React, { Component } from 'react';
 import { ScrollView } from 'react-native';
 import EStyleSheet from 'react-native-extended-stylesheet';
 import Constants from 'expo-constants';
-import ThreeHexagonsWithMargin from '../features/colors/ThreeHexagonsWithMargin';
+import HexagonGrid from '../features/colors/HexagonGrid';
 import concertApi from '../api/concertApi';
 
 class ColorsScreen extends Component {
   handleChoice = color => () => {
-    console.log(`Color choice made. Emitting color choice ${color}`);
-    concertApi.send(
-      JSON.stringify({
-        event: 'CHOICE_MADE',
-        data: {
-          choiceType: 'CHOICE_COLOR',
-          choice: color,
-          timestamp: new Date().toISOString(), // temporary
-        },
-      })
-    );
+    concertApi.send({
+      event: 'CHOICE_MADE',
+      data: {
+        choiceType: 'CHOICE_COLOR',
+        choice: color,
+        timestamp: new Date().toISOString(), // temporary
+      },
+    });
     this.props.navigation.goBack();
   };
 
   render() {
     return (
       <ScrollView horizontal pagingEnabled style={styles.container}>
-        <ThreeHexagonsWithMargin onChoice={this.handleChoice} />
+        <HexagonGrid onChoice={this.handleChoice} />
       </ScrollView>
     );
   }
